+++ packages/atproto/lib/src/services/server_service.dart
@@ -8,6 +8,7 @@ import 'package:atproto_core/atproto_core.dart' as core;
 // 🌎 Project imports:
 import '../nsids.g.dart' as ns;
 import 'entities/account.dart';
+import 'entities/account_status.dart';
 import 'entities/app_password.dart';
 import 'entities/app_passwords.dart';
 import 'entities/created_invite_code.dart';
@@ -217,6 +218,13 @@ final class ServerService {
           'aud': aud,
         },
         to: ServiceAuthToken.fromJson,
+      );
+
+  /// https://atprotodart.com/docs/lexicons/com/atproto/server/checkAccountStatus
+  Future<core.XRPCResponse<AccountStatus>> checkAccountStatus() async =>
+      await _ctx.get(
+        ns.comAtprotoServerCheckAccountStatus,
+        to: AccountStatus.fromJson,
       );
 
   /// https://atprotodart.com/docs/lexicons/com/atproto/server/deactivateAccount